requirements.txt
Django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
channels
python-dotenv